+++ src/core/RUnit.cpp
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2011-2017 by Andrew Mustun. All rights reserved.
+ * Copyright (c) 2011-2018 by Andrew Mustun. All rights reserved.
  * 
  * This file is part of the QCAD project.
  *
@@ -781,11 +781,11 @@ QString RUnit::doubleToString(double value, double prec,
     QString ret;
     QString exaStr;
     int dotPos;
-    long int num;
+    double num;
     if (prec>1.0e-12) {
         // note: msvc does not have round:
         double v = value / prec;
-        num = (v-floor(v)<0.5 ? (long int)floor(v) : (long int)ceil(v));
+        num = (v-floor(v)<0.5 ? floor(v) : ceil(v));
     }
     else {
         num = RMath::mround(value);
